requirement.txt
django>=3.2.0
djangorestframework>=3.9.0,<3.10.0
PyMySQL==0.9.3
django-cors-headers==3.4.0